requirements.txt
mlflow==2.10
fastapi[standard]==0.112.0
uvicorn
pandas~=2.3.1
psycopg2~=2.9.10
psycopg2-binary~=2.9.10
setuptools
wheel
numpy~=1.26.4
statsmodels~=0.14.2
matplotlib~=3.7.5
seaborn~=0.13.2
python-dotenv~=1.1.1
scikit-learn~=1.5.1
mlflow-skinny~=3.2.0
pydantic~=2.11.7
streamlit~=1.37.1
requests~=2.32.3
sqlalchemy~=1.4.54

prefect~=3.4.17